chore(prism-agent): refactor background jobs and split between Connect, Issue, Present, and DID publication state sync
