+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/PrismAgentApp.scala
@@ -3,7 +3,7 @@ package io.iohk.atala.agent.server
 import io.iohk.atala.agent.notification.WebhookPublisher
 import io.iohk.atala.agent.server.config.AppConfig
 import io.iohk.atala.agent.server.http.{ZHttp4sBlazeServer, ZHttpEndpoints}
-import io.iohk.atala.agent.server.jobs.{BackgroundJobs, ConnectBackgroundJobs}
+import io.iohk.atala.agent.server.jobs.{IssueBackgroundJobs, ConnectBackgroundJobs, DIDPublicationBackgroundJobs, PresentBackgroundJobs}
 import io.iohk.atala.agent.walletapi.model.{Entity, Wallet, WalletSeed}
 import io.iohk.atala.agent.walletapi.service.{EntityService, ManagedDIDService, WalletManagementService}
 import io.iohk.atala.castor.controller.{DIDRegistrarServerEndpoints, DIDServerEndpoints}
@@ -56,7 +56,7 @@ object PrismAgentApp {
         .mapError(_.toThrowable)
         .flatMap { wallets =>
           ZIO.foreach(wallets) { wallet =>
-            BackgroundJobs.issueCredentialDidCommExchanges
+            IssueBackgroundJobs.issueCredentialDidCommExchanges
               .provideSomeLayer(ZLayer.succeed(WalletAccessContext(wallet.id))) @@ Metric
               .gauge("issuance_flow_did_com_exchange_job_ms_gauge")
               .trackDurationWith(_.toMetricsSeconds)
@@ -78,7 +78,7 @@ object PrismAgentApp {
         .mapError(_.toThrowable)
         .flatMap { wallets =>
           ZIO.foreach(wallets) { wallet =>
-            BackgroundJobs.presentProofExchanges
+            PresentBackgroundJobs.presentProofExchanges
               .provideSomeLayer(ZLayer.succeed(WalletAccessContext(wallet.id))) @@ Metric
               .gauge("present_proof_flow_did_com_exchange_job_ms_gauge")
               .trackDurationWith(_.toMetricsSeconds)
@@ -114,7 +114,7 @@ object PrismAgentApp {
       .serviceWithZIO[WalletManagementService](_.listWallets().map(_._1))
       .flatMap { wallets =>
         ZIO.foreach(wallets) { wallet =>
-          BackgroundJobs.syncDIDPublicationStateFromDlt
+          DIDPublicationBackgroundJobs.syncDIDPublicationStateFromDlt
             .provideSomeLayer(ZLayer.succeed(WalletAccessContext(wallet.id)))
         }
       }
