+++ prism-agent/service/server/src/main/scala/io/iohk/atala/agent/server/jobs/BackgroundJobsHelper.scala
@@ -1,0 +1,75 @@
+package io.iohk.atala.agent.server.jobs
+
+import io.iohk.atala.agent.walletapi.model.error.DIDSecretStorageError.KeyNotFoundError
+import io.iohk.atala.agent.walletapi.model.{ManagedDIDState, PublicationState}
+import io.iohk.atala.agent.walletapi.service.ManagedDIDService
+import io.iohk.atala.castor.core.model.did.{LongFormPrismDID, PrismDID, VerificationRelationship}
+import io.iohk.atala.castor.core.service.DIDService
+import io.iohk.atala.mercury.model.DidId
+import io.iohk.atala.mercury.{AgentPeerService, DidAgent}
+import io.iohk.atala.pollux.vc.jwt.{ES256KSigner, Issuer as JwtIssuer}
+import io.iohk.atala.shared.models.WalletAccessContext
+import zio.{ZIO, ZLayer}
+
+trait BackgroundJobsHelper {
+
+  def getLongForm(
+      did: PrismDID,
+      allowUnpublishedIssuingDID: Boolean = false
+  ): ZIO[ManagedDIDService & WalletAccessContext, Throwable, LongFormPrismDID] = {
+    for {
+      managedDIDService <- ZIO.service[ManagedDIDService]
+      didState <- managedDIDService
+        .getManagedDIDState(did.asCanonical)
+        .mapError(e => RuntimeException(s"Error occurred while getting did from wallet: ${e.toString}"))
+        .someOrFail(RuntimeException(s"Issuer DID does not exist in the wallet: $did"))
+        .flatMap {
+          case s @ ManagedDIDState(_, _, PublicationState.Published(_)) => ZIO.succeed(s)
+          case s => ZIO.cond(allowUnpublishedIssuingDID, s, RuntimeException(s"Issuer DID must be published: $did"))
+        }
+      longFormPrismDID = PrismDID.buildLongFormFromOperation(didState.createOperation)
+    } yield longFormPrismDID
+  }
+
+  def createJwtIssuer(
+      jwtIssuerDID: PrismDID,
+      verificationRelationship: VerificationRelationship
+  ): ZIO[DIDService & ManagedDIDService & WalletAccessContext, Throwable, JwtIssuer] = {
+    for {
+      managedDIDService <- ZIO.service[ManagedDIDService]
+      didService <- ZIO.service[DIDService]
+      // Automatically infer keyId to use by resolving DID and choose the corresponding VerificationRelationship
+      issuingKeyId <- didService
+        .resolveDID(jwtIssuerDID)
+        .mapError(e => RuntimeException(s"Error occured while resolving Issuing DID during VC creation: ${e.toString}"))
+        .someOrFail(RuntimeException(s"Issuing DID resolution result is not found"))
+        .map { case (_, didData) => didData.publicKeys.find(_.purpose == verificationRelationship).map(_.id) }
+        .someOrFail(
+          RuntimeException(s"Issuing DID doesn't have a key in ${verificationRelationship.name} to use: $jwtIssuerDID")
+        )
+      ecKeyPair <- managedDIDService
+        .javaKeyPairWithDID(jwtIssuerDID.asCanonical, issuingKeyId)
+        .mapError(e => RuntimeException(s"Error occurred while getting issuer key-pair: ${e.toString}"))
+        .someOrFail(
+          RuntimeException(s"Issuer key-pair does not exist in the wallet: ${jwtIssuerDID.toString}#$issuingKeyId")
+        )
+      (privateKey, publicKey) = ecKeyPair
+      jwtIssuer = JwtIssuer(
+        io.iohk.atala.pollux.vc.jwt.DID(jwtIssuerDID.toString),
+        ES256KSigner(privateKey),
+        publicKey
+      )
+    } yield jwtIssuer
+  }
+
+  def buildDIDCommAgent(
+      myDid: DidId
+  ): ZIO[ManagedDIDService & WalletAccessContext, KeyNotFoundError, ZLayer[Any, Nothing, DidAgent]] = {
+    for {
+      managedDidService <- ZIO.service[ManagedDIDService]
+      peerDID <- managedDidService.getPeerDID(myDid)
+      agent = AgentPeerService.makeLayer(peerDID)
+    } yield agent
+  }
+
+}
